+++ apps/frontend/src/app/api/user/reset/route.ts
@@ -1,11 +1,29 @@
-export async function DELETE(req: NextRequest) {
-    const user_id = await getUserId(req);
+import { ApiError, ApiResponse } from "@/utils/apiResponse";
+import { getUserIdFromRequest } from "@/utils/auth";
+import { prisma } from "@repo/db";
+import { NextRequest, NextResponse } from "next/server";
 
-    await prisma.$transaction([
-        prisma.history.deleteMany({ where: { user_id } }),
-        prisma.totalStatistics.updateMany({ where: { user_id }, data: { total_letters_typed: 0, total_tests_taken: 0, total_words_typed: 0 } }),
-        prisma.leaderboard.deleteMany({ where: { user_id } }),
-    ]);
+export async function POST(req: NextRequest) {
 
-    return NextResponse.json(new ApiResponse(200, "Account stats reset successfully"), { status: 200 });
+    try {
+        const user_id = await getUserIdFromRequest(req);
+
+        if (!user_id) {
+            return NextResponse.json(new ApiError(401, "Unauthorized"), {
+                status: 401,
+            });
+        }
+
+
+        await prisma.$transaction([
+            prisma.history.deleteMany({ where: { user_id } }),
+            prisma.totalStatistics.updateMany({ where: { user_id }, data: { total_letters_typed: 0, total_tests_taken: 0, total_words_typed: 0 } }),
+            prisma.leaderboard.deleteMany({ where: { user_id } }),
+        ]);
+
+        return NextResponse.json(new ApiResponse(200, null, "Account stats reset successfully"), { status: 200 });
+
+    } catch {
+        return NextResponse.json(new ApiError(500, "Something went wrong while reseting account"))
+    }
 }
